add badges function to data.js
