+++ src/user/data.js
@@ -87,6 +87,47 @@ module.exports = function (User) {
         "email:confirmed": 0,
     };
 
+    // type signature:
+    // interface User {
+    //    calculateBadge(uid: number): Promise<string>;
+    // }
+    User.calculateBadge = async function (uid) {
+
+        if (typeof uid !== 'number') {
+            throw new TypeError('uid must be a number');
+        }
+
+        const reputation = await user.getUserField(uid, 'reputation');
+        const postCount = await user.getUserField(uid, 'postcount');
+        const userBadges = [];
+
+        // users can have multiple badges based on
+        // reputation and post count statistics
+        if (reputation > 5) {
+            userBadges.push('⭐');
+        } else if (reputation > 20) {
+            userBadges.push('🌟');
+        } else {
+            userBadges.push('💫');
+        }
+
+        if (postCount > 5) {
+            userBadges.push('🌱');
+        } else if (postCount > 20) {
+            userBadges.push('🌷');
+        } else {
+            userBadges.push('🌳');
+        }
+
+        result = userBadges.join('');
+
+        if (typeof result !== 'string') {
+            throw new TypeError('The function must return a string');
+        }
+
+        return result;
+    };
+
     User.getUsersFields = async function (uids, fields) {
         if (!Array.isArray(uids) || !uids.length) {
             return [];
